refactor(layout): move apply reconciler into vmux_layout::reconcile
