+++ crates/vmux_desktop/src/agent_layout/reconcile.rs
@@ -1,1 +1,2 @@
+#[allow(unused_imports)]
 pub use vmux_layout::reconcile::{ValidationError, plan_diff};
